requirements.txt
# 文档处理
python-docx>=0.8.11
pywin32>=305

# PDF处理
pypdf2>=3.0.0
pdfplumber>=0.9.0

# arXiv API
arxiv>=1.4.7

# HTTP请求
requests>=2.28.0

# 进度条
tqdm>=4.65.0

# DeepSeek API
openai>=1.0.0

# 其他工具
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0